pyproject.toml
[project]
name = "langgraph-learning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.5",
    "langchain>=0.3.26",
    "langchain-google-genai>=2.0.10",
    "langgraph>=0.5.0",
    "python-dotenv>=1.1.1",
]
